@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) Pieter Svenson
+ * Copyright (c) whimxiqal
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,7 +23,7 @@
 
 package net.whimxiqal.journey.bukkit.listener;
 
-import net.whimxiqal.journey.common.Journey;
+import net.whimxiqal.journey.Journey;
 import net.whimxiqal.journey.bukkit.util.BukkitUtil;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
