@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) Pieter Svenson
+ * Copyright (c) whimxiqal
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,8 +23,8 @@
 
 package net.whimxiqal.journey.bukkit.navigation.mode;
 
-import net.whimxiqal.journey.common.navigation.Mode;
-import net.whimxiqal.journey.common.search.SearchSession;
+import net.whimxiqal.journey.navigation.Mode;
+import net.whimxiqal.journey.search.SearchSession;
 import net.whimxiqal.journey.bukkit.util.BukkitUtil;
 import java.util.Set;
 import org.bukkit.Material;
