@@ -31,8 +31,9 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Scanner;
-import net.whimxiqal.journey.common.navigation.Cell;
-import net.whimxiqal.journey.common.navigation.ModeType;
+import net.whimxiqal.journey.Cell;
+import net.whimxiqal.journey.Journey;
+import net.whimxiqal.journey.Tunnel;
 
 public final class WorldLoader {
 
@@ -45,10 +46,16 @@ public final class WorldLoader {
     // all the rest are "blank" but serve some other purpose too
   }
 
+  public static int domain(int index) {
+    return Journey.get().domainManager().domainIndex(worldResources[index]);
+  }
+
   public static void initWorlds() {
     Map<Character, Cell> pendingPorts = new HashMap<>();
 
-    for (String resource : worldResources) {
+    for (int resourceIdx = 0; resourceIdx < worldResources.length; resourceIdx++) {
+      String resource = worldResources[resourceIdx];
+      int domain = domain(resourceIdx);
       TestWorld world = new TestWorld();
       world.name = resource;
 
@@ -69,30 +76,24 @@ public final class WorldLoader {
 
             // POI
             if (Character.isDigit(c)) {
-              TestPlatformProxy.pois.put(Character.toString(c), new Cell(x, y, 0, resource));
+              TestPlatformProxy.pois.put(Character.toString(c), new Cell(x, y, 0, domain));
             }
 
-            // PORT
+            // TUNNEL
             if (Character.isLetter(c)) {
-              Cell cell = new Cell(x, y, 0, resource);
+              Cell cell = new Cell(x, y, 0, domain);
               if (Character.isUpperCase(c)) {
                 if (pendingPorts.containsKey(Character.toLowerCase(c))) {
-                  // complete the port
-                  TestPlatformProxy.ports.add(new TestPort(pendingPorts.get(Character.toLowerCase(c)),
-                      cell,
-                      ModeType.PORT,
-                      1));
+                  // complete the tunnel
+                  TestPlatformProxy.tunnels.add(Tunnel.builder(pendingPorts.get(Character.toLowerCase(c)), cell).build());
                 } else {
                   pendingPorts.put(c, cell);
                 }
               } else {
                 // lower case
                 if (pendingPorts.containsKey(Character.toUpperCase(c))) {
-                  // complete the port
-                  TestPlatformProxy.ports.add(new TestPort(cell,
-                      pendingPorts.get(Character.toUpperCase(c)),
-                      ModeType.PORT,
-                      1));
+                  // complete the tunnel
+                  TestPlatformProxy.tunnels.add(Tunnel.builder(cell, pendingPorts.get(Character.toUpperCase(c))).build());
                 } else {
                   pendingPorts.put(c, cell);
                 }
@@ -117,7 +118,7 @@ public final class WorldLoader {
         }
       }
 
-      TestPlatformProxy.worlds.put(world.name, world);
+      TestPlatformProxy.worlds.put(domain, world);
     }
   }
 
